Separate CartImage to own component
